requirements.txt
Django==1.11.18
django-extensions==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
django-suit==0.2.25
django-suit-daterange-filter==0.0.4
django-nested-admin==3.0.19
django-tinymce==2.6.0
Pillow==4.0.0
django-wysiwyg==0.8.0
django-ckeditor==5.3.0
django-modeltranslation==0.12.2
ipython==7.2.0
django-hashid-field==2.1.5
hashids==1.2.0
djangorestframework-jwt==1.11.0
django-suit==0.2.25
django-suit-daterange-filter==0.0.4
django-nested-admin==3.0.19
django-tinymce==2.6.0
Pillow==4.0.0
coreapi==2.3.3